code/backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-multipart==0.0.6
pydantic==2.9.2
python-dotenv==1.0.1
cloudinary==1.41.0
python-socketio==5.11.4
PyJWT==2.9.0
email-validator==2.2.0
